requirements.txt
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.1.post1
python-dotenv==1.0.0
tqdm==4.66.1
pulp==2.8.0
google-api-python-client==2.120.0
textblob==0.18.0
nltk==3.8.1 